Improved HTML, CSS and JS
