+++ app.css
@@ -28,7 +28,7 @@ a{
     color: #fff
 }
 
-div.lectures, div#weekDay{
+div.lectures, header{
     cursor: default;
     font-family: var(--fontStyle);
     border: 1px solid #e0e0e0;
@@ -41,7 +41,10 @@ div.lectures p:last-child{
     border: none
 }
 
-div#weekDay{
+header{
+    /* padding: 10px; */
+    margin-left: 15%;
+    margin-right: 15%;
     text-align: center;
     border: none;
     border-radius: 19px;
@@ -120,9 +123,11 @@ button {
 .arrow:hover polyline,.arrow:active polyline{
     stroke-width: 6px
 }
-div.nav{
+nav{
     display: flex;
-    justify-content: space-between
+    justify-content: space-between;
+    position: relative;
+    top: -2.2em;
 }
 /* Arrow CSS End */
 
@@ -131,5 +136,19 @@ div.nav{
     footer{
         margin: auto auto;
         width: 400px
+    }
+}
+
+@media only screen and (min-width: 540px){
+    nav{
+        margin-left: 5%;
+        margin-right: 5%;
+    }
+}
+
+@media only screen and (min-width: 746px){
+    nav{
+        margin-left: 10%;
+        margin-right: 10%;
     }
 }
