+++ js/DOMFunctions.js
@@ -2,19 +2,13 @@ import foo from './functions.js';
 
 function weekDayDOM(weekDay){
     // Selecting a mainDivBody element which is 'div.lectures'
-    let mainDivBody = document.querySelector('div#weekDay');
-
-    // Creating 'p' tag element
-    let pTag = document.createElement('p');
-    pTag.className = 'lecItem weekDay';
-    pTag.textContent = `${weekDay} ${foo.shuffle(['😄','😍','😅','😎','😥','🙄',
+    let mainPTag = document.querySelector('header p.weekDay');
+    mainPTag.textContent = `${weekDay} ${foo.shuffle(['😄','😍','😅','😎','😥','🙄',
                                           '😫','😪','🤐','😴','🥱','😝',
                                           '😒','😕','🙃','😲','😖','😤',
                                           '🤯','🤪','😵','🤬','🥺','😇',
                                           '🤫','🤭','🧐','😈','🤳','🙏',
                                           '💻','🔞','💔',])[0]}`;
-
-    mainDivBody.appendChild(pTag);
 }
 
 function aTagDOM(theClass,link,authuser){
@@ -30,7 +24,6 @@ function aTagDOM(theClass,link,authuser){
 function spanTag(time){
     //Creating 'span' Tag for timing of the lecture/period
     let spanTag = document.createElement('span');
-    // ${timing[lecCount]}
     spanTag.textContent = `{${time}}`;
 
     return spanTag;
@@ -57,13 +50,37 @@ function basePTagOfMultipleClass(spanTag,lectureNumber){
     return basePTag;
 }
 
-function DOMReset(logo){
+function DOMReset(){
     // it empties the 'div.lectures' tag/element
     document.querySelector('div.lectures').innerHTML='';
-    //'div#weekDay' tag/element's content overrided by logo element 
-    document.querySelector('div#weekDay').innerHTML = logo.innerHTML;
+}
+
+function videoHide(){
+    let vidDiv = document.querySelector('div.videoDiv');
+    vidDiv.style.visibility = "hidden";
+    vidDiv.style.height = "0";
+}
+
+//It is DOM Manipulation fuction for Video Element
+function setVideo(){
+    let videoBanner = document.querySelector('video.video-banner');
+    // empties the videoBanner element
+    videoBanner.pause();
+    videoBanner.innerHTML = "";
+    videoBanner.load();
+
+    let sourceVid = document.createElement('source');
+
+    // Generate Random no between 1-8
+    let num = ((Math.random()*7)+1);
+    num = Math.round(num);
+
+    sourceVid.setAttribute("src",`./assets/${num}.mp4`);
+    sourceVid.setAttribute("type","video/mp4");
+    sourceVid.textContent = "Your browser does not support the video tag.";
+    videoBanner.appendChild(sourceVid);
 }
 
 export default{
-    weekDayDOM, aTagDOM, spanTag, basePTag, basePTagOfMultipleClass, DOMReset
+    weekDayDOM, aTagDOM, spanTag, basePTag, basePTagOfMultipleClass, DOMReset, videoHide, setVideo
 };
